prevent multiple sublists showing at once
